Safari-611.1.10/LayoutTests/storage/domstorage/sessionstorage/blocked-file-access-expected.txt

CONSOLE MESSAGE: Not allowed to load local resource: clearSessionStorage.js
CONSOLE MESSAGE: PASS: window.sessionStorage was accessible

Test that we are permitted access to sessionStorage from a file URL if universal access is turned off.
